Clean up HubMain
